Fixed it didn't compile for Windows CE.
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@4362 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qs/src/ui/window.cpp
@@ -219,8 +219,8 @@ bool qs::Window::centerWindow(HWND hwnd)
 	RECT rect;
 	getWindowRect(&rect);
 	
-	int nLeft = (rectParent.left + rectParent.right - (rect.right - rect.left))/2;
-	int nTop = (rectParent.top + rectParent.bottom - (rect.bottom - rect.top))/2;
+	long nLeft = (rectParent.left + rectParent.right - (rect.right - rect.left))/2;
+	long nTop = (rectParent.top + rectParent.bottom - (rect.bottom - rect.top))/2;
 
 	RECT rectMonitor;
 #if !defined _WIN32_WCE && WINVER >= 0x500
